@@ -1,8 +1,8 @@
 // @ts-check
 // Note: type annotations allow type checking and IDEs autocompletion
 
-const lightCodeTheme = require('prism-react-renderer/themes/github');
-const darkCodeTheme = require('prism-react-renderer/themes/dracula');
+import {themes as prismThemes} from 'prism-react-renderer';
+
 const DefaultLocale = 'en';
 
 /** @type {import('@docusaurus/types').Config} */
@@ -100,8 +100,8 @@ const config = {
         copyright: `Copyright © ${new Date().getFullYear()} Famedly GmbH`,
       },
       prism: {
-        theme: lightCodeTheme,
-        darkTheme: darkCodeTheme,
+        theme: prismThemes.github,
+        darkTheme: prismThemes.dracula,
       },
       algolia: {
         appId: 'FFJ2EFU29K',
